Make KeyValue configuration available for any filter (disassociate from AddRequestHeadersIfNotPresentGatewayFilterFactory)
