+++ spring-cloud-gateway-server/src/main/java/org/springframework/cloud/gateway/support/config/KeyValue.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2013-2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.gateway.support.config;
+
+import org.springframework.core.style.ToStringCreator;
+
+/**
+ * @author Marta Medio
+ */
+public class KeyValue {
+
+	private final String key;
+
+	private final String value;
+
+	public KeyValue(String key, String value) {
+		this.key = key;
+		this.value = value;
+	}
+
+	public String getKey() {
+		return key;
+	}
+
+	public String getValue() {
+		return value;
+	}
+
+	@Override
+	public String toString() {
+		return new ToStringCreator(this).append("name", key).append("value", value).toString();
+	}
+
+}
