+++ spring-cloud-gateway-server/src/main/java/org/springframework/cloud/gateway/support/config/KeyValueConfig.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2013-2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.gateway.support.config;
+
+/**
+ * Adds a new format of configuration for filters. - Filter: key:value,key1:value1
+ *
+ * @author Marta Medio
+ */
+public class KeyValueConfig {
+
+	private KeyValue[] keyValues;
+
+	public KeyValue[] getKeyValues() {
+		return keyValues;
+	}
+
+	public void setKeyValues(KeyValue[] keyValues) {
+		this.keyValues = keyValues;
+	}
+
+}
